fix(ios): real-time messaging UX, audio accent colors, composer language priority
- Scroll: use .defaultScrollAnchor(.bottom) for instant WhatsApp-style opening
- Scroll: target bottom_spacer on new messages to keep content above composer
- Background refresh: increment newMessageAppended to trigger auto-scroll
- Audio: sent audio uses MeeshyColors.brandPrimaryHex, received uses otherBubbleColor
- Bubbles: replace hardcoded hex strings with MeeshyColors SDK constants
- Language: keyboard layout > system language > "fr" fallback
- TextAnalyzer: lock language detection after 10 words instead of confidence-only
- WaveformCache: fix Int16.min overflow crash in abs() via Int32 cast

+++ packages/MeeshySDK/Sources/MeeshySDK/Audio/WaveformCache.swift
@@ -137,7 +137,7 @@ public actor WaveformCache {
                 let int16Ptr = ptr.bindMemory(to: Int16.self)
                 for sample in int16Ptr {
                     let bucketIndex = min(globalSampleIndex / estimatedBucketSize, count - 1)
-                    buckets[bucketIndex] += Float(abs(sample)) / Float(Int16.max)
+                    buckets[bucketIndex] += Float(abs(Int32(sample))) / Float(Int16.max)
                     bucketCounts[bucketIndex] += 1
                     globalSampleIndex += 1
                 }
